+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@ Release date: unreleased
 
 - Allows aiosmtpd version 1.3.2.
 - `SMPTDFix` can now be initialized without providing a hostname and port which defaults to localhost: 8025.
+- Better compatibility with AUTH in aiosmtpd by using `AuthResult` as return from AUTH mechanism handlers and the `Authenticator` for processing.
 
 ## Version 0.2.9